Rename htr2hpc.train.apps to htr2hpc.train.data
